fix : 경고 메시지 제거, Mushroom아이템 이동 방식 수정 (#20)
* fix : Gunba클래스의 오류 메시지 제거

* fix : Mushroom클래스의 경고메시지 일부 제거

* fix : FireAmmo의 경고 메시지 일부 제거

* fix : TempScene의 경고 메시지 제거

* fix : Item, MyTileMapTool, playerCharacter클래스 오류 메시지 제거

* feat : Mushroom아이템 이동 방식 수정

플레이어 이동에 따라 움직이도록 변경

+++ Gunba.cpp
@@ -68,10 +68,10 @@ void Gunba::ChangeAnimationFrame()
 
 void Gunba::UpdateCollider()
 {
-    SetRect(&collider, pos.x - img->GetFrameWidth() / 2,
-        pos.y - img->GetFrameWidth() / 2,
-        pos.x + img->GetFrameWidth() / 2,
-        pos.y + img->GetFrameHeight() / 2);
+    SetRect(&collider, (int)pos.x - img->GetFrameWidth() / 2,
+        (int)pos.y - img->GetFrameWidth() / 2,
+        (int)pos.x + img->GetFrameWidth() / 2,
+        (int)pos.y + img->GetFrameHeight() / 2);
 }
 
 void Gunba::UpdatePosition()
@@ -210,8 +210,8 @@ void Gunba::Update()
 
     ChangeDirection();
 
-    nowTileIndexX = (pos.x + GLOBAL_POS) / INGAME_RENDER_TILE_WIDHT_COUNT;
-    nowTileIndexY = pos.y / MAP_HEIGHT;
+    nowTileIndexX = (int)(pos.x + GLOBAL_POS) / INGAME_RENDER_TILE_WIDHT_COUNT;
+    nowTileIndexY = (int)pos.y / MAP_HEIGHT;
 
     CheckOutWindow();
 }
@@ -221,9 +221,9 @@ void Gunba::Render(HDC hdc)
     if (isDead == true)
         return;
 
-    Rectangle(hdc, TILE_DATA[nowTileIndexY - 1][nowTileIndexX].rc.left - GLOBAL_POS,
+    Rectangle(hdc, TILE_DATA[nowTileIndexY - 1][nowTileIndexX].rc.left - (int)GLOBAL_POS,
         TILE_DATA[nowTileIndexY - 1][nowTileIndexX].rc.top,
-        TILE_DATA[nowTileIndexY - 1][nowTileIndexX].rc.right - GLOBAL_POS,
+        TILE_DATA[nowTileIndexY - 1][nowTileIndexX].rc.right - (int)GLOBAL_POS,
         TILE_DATA[nowTileIndexY - 1][nowTileIndexX].rc.bottom);
 
     img->Render(hdc, pos.x, pos.y, frameX, frameY);
